refactor(reader): put the settings a reader never changes behind Advanced
The typography sheet was meant to stay the handful of answers a reader
changes often, with everything rarer a tap further in. It grew to eleven
controls anyway, one reasonable row at a time, while the Advanced sheet
it was supposed to feed held exactly one.

Move six of them across: line height, page margins, columns, the footer
mode, the page-turn animation and the just-this-book toggle. What is
left is the theme, the size, the light, the face, how the book is read
and whether the screen stays awake for it.

The Advanced row stops being conditional. It was hidden in a paginated
book because auto-scroll was all that lived behind it and the way in to
an empty sheet is worse than no way in; with six more rows there it
cannot be empty. The rows that do not apply still hide themselves.

Settings shows the same controls without a book, so it carries the same
split behind a collapsed Advanced section, and AGENTS.md now says a new
reading setting lands in Advanced unless there is a case for the first
sheet — which is what stops this happening again.
