+++ app/src/main/kotlin/com/chmouel/liseur/reader/chrome/AdvancedSheet.kt
@@ -1,5 +1,6 @@
 package com.chmouel.liseur.reader.chrome
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -27,7 +28,13 @@ import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.dp
 import com.chmouel.liseur.R
 import com.chmouel.liseur.data.settings.AutoScrollPreference
+import com.chmouel.liseur.data.settings.ColumnMode
+import com.chmouel.liseur.data.settings.FooterMode
+import com.chmouel.liseur.data.settings.ReaderPrefs
 import com.chmouel.liseur.ui.contentWidthCap
+import com.chmouel.liseur.ui.reading.ReadingFooterModeDropdown
+import com.chmouel.liseur.ui.reading.ReadingLayoutControls
+import com.chmouel.liseur.ui.reading.ReadingPageTurnAnimationToggle
 import com.chmouel.liseur.ui.reading.ReadingSectionLabel
 import com.chmouel.liseur.ui.windowWidth
 
@@ -35,25 +42,38 @@ import com.chmouel.liseur.ui.windowWidth
  * The second sheet, reached from the typography sheet's last row.
  *
  * It exists so that the typography sheet does not have to grow: that one
- * stays the four or five answers a reader changes often, and everything
+ * stays the handful of answers a reader changes often — the theme, the
+ * size, the light, the face, and how the book is read — and everything
  * rarer lives a tap further in. Dismissing this lands back on
  * typography, and dismissing that lands back on the page.
  *
- * Today it holds one row. The rest of what belongs here — finer
- * typography, read-aloud, imported fonts, tap-zone presets — arrives with
- * its own issue, and the shape of the sheet is the point: one row, or
- * five, it is the same sheet and the reader learns it once.
+ * What it holds is what a reader sets once, if ever: the shape of the
+ * text block, what the footer says, whether pages turn with an
+ * animation, whether the page moves on its own, and whether any of it is
+ * this book's alone. The rest of what belongs here — finer typography,
+ * read-aloud, imported fonts, tap-zone presets — arrives with its own
+ * issue, and the shape of the sheet is the point: five rows, or ten, it
+ * is the same sheet and the reader learns it once.
  *
  * See `docs/adr/0001-advanced-reading-menu.md`.
  */
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun AdvancedSheet(
+    prefs: ReaderPrefs,
+    scrollMode: Boolean,
     autoScrollOffered: Boolean,
     autoScrolling: Boolean,
     autoScrollSpeed: Float,
+    typographyIsOwn: Boolean,
+    onLineHeightChanged: (Double?) -> Unit,
+    onPageMarginsChanged: (Double?) -> Unit,
+    onColumnModeChanged: (ColumnMode) -> Unit,
+    onFooterModeChanged: (FooterMode) -> Unit,
+    onPageTurnAnimationChanged: (Boolean) -> Unit,
     onAutoScrollChanged: (Boolean) -> Unit,
     onAutoScrollSpeedChanged: (Float) -> Unit,
+    onTypographyIsOwnChanged: (Boolean) -> Unit,
     onDismiss: () -> Unit,
 ) {
     ModalBottomSheet(onDismissRequest = onDismiss) {
@@ -66,6 +86,30 @@ fun AdvancedSheet(
                 .padding(bottom = 24.dp),
             verticalArrangement = Arrangement.spacedBy(20.dp),
         ) {
+            ReadingLayoutControls(
+                lineHeight = prefs.lineHeight,
+                pageMargins = prefs.pageMargins,
+                columnMode = prefs.columnMode,
+                // A scrolled chapter is one running column, so the count
+                // has nothing to divide and the control would take a tap
+                // and change nothing.
+                showColumns = !scrollMode,
+                onLineHeightChanged = onLineHeightChanged,
+                onPageMarginsChanged = onPageMarginsChanged,
+                onColumnModeChanged = onColumnModeChanged,
+            )
+            ReadingFooterModeDropdown(
+                selected = prefs.footerMode,
+                onSelected = onFooterModeChanged,
+            )
+            // Nothing lifts off the screen when the text is scrolled, so
+            // the animation has no page to describe.
+            if (!scrollMode) {
+                ReadingPageTurnAnimationToggle(
+                    enabled = prefs.pageTurnAnimation,
+                    onChanged = onPageTurnAnimationChanged,
+                )
+            }
             if (autoScrollOffered) {
                 AutoScrollRow(
                     enabled = autoScrolling,
@@ -74,7 +118,46 @@ fun AdvancedSheet(
                     onSpeedChanged = onAutoScrollSpeedChanged,
                 )
             }
+            JustThisBookToggle(
+                enabled = typographyIsOwn,
+                onChanged = onTypographyIsOwnChanged,
+            )
         }
+    }
+}
+
+/**
+ * Sets the book apart from the shared reading settings.
+ *
+ * Last in the sheet, and only about the settings that are genuinely the
+ * book's: what it is set in, how big, how open, how wide. The theme and
+ * the brightness are about the room, so they carry on being shared
+ * however this is left.
+ */
+@Composable
+private fun JustThisBookToggle(enabled: Boolean, onChanged: (Boolean) -> Unit) {
+    Row(
+        verticalAlignment = Alignment.CenterVertically,
+        modifier = Modifier
+            .fillMaxWidth()
+            .clickable { onChanged(!enabled) },
+    ) {
+        Column(Modifier.weight(1f)) {
+            Text(
+                text = stringResource(R.string.reader_typography_own),
+                style = MaterialTheme.typography.bodyLarge,
+            )
+            Text(
+                text = if (enabled) {
+                    stringResource(R.string.reader_typography_own_on)
+                } else {
+                    stringResource(R.string.reader_typography_own_off)
+                },
+                style = MaterialTheme.typography.bodyMedium,
+                color = MaterialTheme.colorScheme.onSurfaceVariant,
+            )
+        }
+        Switch(checked = enabled, onCheckedChange = onChanged)
     }
 }
