+++ app/src/main/kotlin/com/chmouel/liseur/reader/chrome/TypographySheet.kt
@@ -25,8 +25,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.unit.dp
 import com.chmouel.liseur.R
-import com.chmouel.liseur.data.settings.ColumnMode
-import com.chmouel.liseur.data.settings.FooterMode
 import com.chmouel.liseur.data.settings.ReaderFont
 import com.chmouel.liseur.data.settings.ReaderPrefs
 import com.chmouel.liseur.data.settings.ReaderTheme
@@ -35,42 +33,36 @@ import com.chmouel.liseur.ui.contentWidthCap
 import com.chmouel.liseur.ui.reading.ReadingBrightnessSlider
 import com.chmouel.liseur.ui.reading.ReadingFontDropdown
 import com.chmouel.liseur.ui.reading.ReadingFontSizeSlider
-import com.chmouel.liseur.ui.reading.ReadingFooterModeDropdown
-import com.chmouel.liseur.ui.reading.ReadingLayoutControls
-import com.chmouel.liseur.ui.reading.ReadingPageTurnAnimationToggle
 import com.chmouel.liseur.ui.reading.ReadingThemeRow
 import com.chmouel.liseur.ui.windowWidth
 
 /**
- * Kindle-style "Aa" sheet: reading theme, font, size, brightness
- * and layout, applied live as they change.
+ * Kindle-style "Aa" sheet: reading theme, font, size and brightness,
+ * applied live as they change.
  *
  * The controls themselves live in `ui/reading`, because Settings shows
  * the same ones on a screen of its own. This sheet is the quick path to
- * them with a book open, and adds the three answers that only make
- * sense with one: scrolling, the screen, and setting this book apart.
+ * them with a book open, and adds the two answers that only make sense
+ * with one: how the book is read, and whether the screen stays awake for
+ * it.
+ *
+ * It is deliberately short. Everything a reader sets once, if ever, is a
+ * tap further in, behind the Advanced row at the bottom — see
+ * `docs/adr/0001-advanced-reading-menu.md`.
  */
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun TypographySheet(
     prefs: ReaderPrefs,
     readingTheme: ReaderTheme,
-    typographyIsOwn: Boolean,
-    onTypographyIsOwnChanged: (Boolean) -> Unit,
     onFontSelected: (ReaderFont) -> Unit,
     onFontSizeChanged: (Double) -> Unit,
     onThemeSelected: (ReaderThemeChoice) -> Unit,
-    onLineHeightChanged: (Double?) -> Unit,
-    onPageMarginsChanged: (Double?) -> Unit,
     onBrightnessChanged: (Float?) -> Unit,
-    onPageTurnAnimationChanged: (Boolean) -> Unit,
-    onFooterModeChanged: (FooterMode) -> Unit,
-    onColumnModeChanged: (ColumnMode) -> Unit,
     keepScreenOn: Boolean,
     onKeepScreenOnChanged: (Boolean) -> Unit,
     scrollMode: Boolean,
     onScrollModeChanged: (Boolean) -> Unit,
-    advancedOffered: Boolean,
     onOpenAdvanced: () -> Unit,
     onDismiss: () -> Unit,
 ) {
@@ -92,48 +84,17 @@ fun TypographySheet(
             ReadingFontSizeSlider(value = prefs.fontSize, onChanged = onFontSizeChanged)
             ReadingBrightnessSlider(value = prefs.brightness, onChanged = onBrightnessChanged)
             ReadingFontDropdown(selected = prefs.font, onSelected = onFontSelected)
-            ReadingLayoutControls(
-                lineHeight = prefs.lineHeight,
-                pageMargins = prefs.pageMargins,
-                columnMode = prefs.columnMode,
-                // A scrolled chapter is one running column, so the count
-                // has nothing to divide and the control would take a tap
-                // and change nothing.
-                showColumns = !scrollMode,
-                onLineHeightChanged = onLineHeightChanged,
-                onPageMarginsChanged = onPageMarginsChanged,
-                onColumnModeChanged = onColumnModeChanged,
-            )
-            ReadingFooterModeDropdown(
-                selected = prefs.footerMode,
-                onSelected = onFooterModeChanged,
-            )
             ScrollModeToggle(
                 enabled = scrollMode,
                 onChanged = onScrollModeChanged,
             )
-            // Nothing lifts off the screen when the text is scrolled, so
-            // the animation has no page to describe.
-            if (!scrollMode) {
-                ReadingPageTurnAnimationToggle(
-                    enabled = prefs.pageTurnAnimation,
-                    onChanged = onPageTurnAnimationChanged,
-                )
-            }
             KeepScreenOnToggle(
                 enabled = keepScreenOn,
                 onChanged = onKeepScreenOnChanged,
             )
-            JustThisBookToggle(
-                enabled = typographyIsOwn,
-                onChanged = onTypographyIsOwnChanged,
-            )
-            // Last, and only when there is something behind it. An
-            // Advanced row that opens an empty sheet is worse than no
-            // row at all.
-            if (advancedOffered) {
-                AdvancedRow(onClick = onOpenAdvanced)
-            }
+            // Last, and always: what is behind it applies to every book,
+            // read either way, so it can no longer open an empty sheet.
+            AdvancedRow(onClick = onOpenAdvanced)
         }
     }
 }
@@ -226,40 +187,5 @@ private fun KeepScreenOnToggle(enabled: Boolean, onChanged: (Boolean) -> Unit) {
             )
         }
         Switch(checked = enabled, onCheckedChange = null)
-    }
-}
-
-/**
- * Sets the book apart from the shared reading settings.
- *
- * Last in the sheet, and only about the four settings above it that are
- * genuinely the book's: what it is set in, how big, how open, how wide.
- * The theme and the brightness are about the room, so they carry on
- * being shared however this is left.
- */
-@Composable
-private fun JustThisBookToggle(enabled: Boolean, onChanged: (Boolean) -> Unit) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable { onChanged(!enabled) },
-    ) {
-        Column(Modifier.weight(1f)) {
-            Text(
-                text = stringResource(R.string.reader_typography_own),
-                style = MaterialTheme.typography.bodyLarge,
-            )
-            Text(
-                text = if (enabled) {
-                    stringResource(R.string.reader_typography_own_on)
-                } else {
-                    stringResource(R.string.reader_typography_own_off)
-                },
-                style = MaterialTheme.typography.bodyMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-        Switch(checked = enabled, onCheckedChange = onChanged)
     }
 }
