+++ app/src/main/kotlin/com/chmouel/liseur/ui/settings/ReadingAppearanceScreen.kt
@@ -1,5 +1,12 @@
 package com.chmouel.liseur.ui.settings
 
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.animation.EnterTransition
+import androidx.compose.animation.ExitTransition
+import androidx.compose.animation.expandVertically
+import androidx.compose.animation.fadeIn
+import androidx.compose.animation.fadeOut
+import androidx.compose.animation.shrinkVertically
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -20,9 +27,14 @@ import androidx.compose.material3.LargeTopAppBar
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -39,6 +51,7 @@ import com.chmouel.liseur.data.settings.ReaderFont
 import com.chmouel.liseur.data.settings.ReaderPrefs
 import com.chmouel.liseur.data.settings.ReaderTheme
 import com.chmouel.liseur.data.settings.ReaderThemeChoice
+import com.chmouel.liseur.ui.LocalEInk
 import com.chmouel.liseur.ui.contentWidthCap
 import com.chmouel.liseur.ui.reading.ReadingBrightnessSlider
 import com.chmouel.liseur.ui.reading.ReadingFontDropdown
@@ -119,25 +132,13 @@ fun ReadingAppearanceScreen(
                 ReadingFontSizeSlider(value = prefs.fontSize, onChanged = onFontSize)
                 ReadingBrightnessSlider(value = prefs.brightness, onChanged = onBrightness)
                 ReadingFontDropdown(selected = prefs.font, onSelected = onFont)
-                ReadingLayoutControls(
-                    lineHeight = prefs.lineHeight,
-                    pageMargins = prefs.pageMargins,
-                    columnMode = prefs.columnMode,
-                    // Unlike the sheet, there is no book here to be
-                    // scrolled, so the count always has something to
-                    // divide and the control is always worth offering.
-                    showColumns = true,
-                    onLineHeightChanged = onLineHeight,
-                    onPageMarginsChanged = onPageMargins,
-                    onColumnModeChanged = onColumnMode,
-                )
-                ReadingFooterModeDropdown(
-                    selected = prefs.footerMode,
-                    onSelected = onFooterMode,
-                )
-                ReadingPageTurnAnimationToggle(
-                    enabled = prefs.pageTurnAnimation,
-                    onChanged = onPageTurnAnimation,
+                AdvancedSection(
+                    prefs = prefs,
+                    onLineHeight = onLineHeight,
+                    onPageMargins = onPageMargins,
+                    onColumnMode = onColumnMode,
+                    onFooterMode = onFooterMode,
+                    onPageTurnAnimation = onPageTurnAnimation,
                 )
                 Text(
                     text = stringResource(R.string.settings_reading_appearance_detail),
@@ -145,6 +146,66 @@ fun ReadingAppearanceScreen(
                     color = MaterialTheme.colorScheme.onSurfaceVariant,
                 )
             }
+        }
+    }
+}
+
+/**
+ * The same split the reader sees, on the screen that has no book.
+ *
+ * What is behind the "Aa" sheet's Advanced row is behind this too, in
+ * the same order: a reader who learned where the margins live in one
+ * place should not have to learn it again in the other. See
+ * `docs/adr/0001-advanced-reading-menu.md`.
+ */
+@Composable
+private fun AdvancedSection(
+    prefs: ReaderPrefs,
+    onLineHeight: (Double?) -> Unit,
+    onPageMargins: (Double?) -> Unit,
+    onColumnMode: (ColumnMode) -> Unit,
+    onFooterMode: (FooterMode) -> Unit,
+    onPageTurnAnimation: (Boolean) -> Unit,
+) {
+    var expanded by rememberSaveable { mutableStateOf(false) }
+
+    TextButton(onClick = { expanded = !expanded }) {
+        Text(
+            stringResource(
+                if (expanded) R.string.reading_hide_advanced else R.string.reading_show_advanced,
+            ),
+        )
+    }
+    // Sliding this open on electronic paper is a full repaint of the
+    // lower half of the screen for every frame; it is the same list
+    // either way, so it simply appears.
+    val eInk = LocalEInk.current
+    AnimatedVisibility(
+        visible = expanded,
+        enter = if (eInk) EnterTransition.None else fadeIn() + expandVertically(),
+        exit = if (eInk) ExitTransition.None else fadeOut() + shrinkVertically(),
+    ) {
+        Column(verticalArrangement = Arrangement.spacedBy(20.dp)) {
+            ReadingLayoutControls(
+                lineHeight = prefs.lineHeight,
+                pageMargins = prefs.pageMargins,
+                columnMode = prefs.columnMode,
+                // Unlike the sheet, there is no book here to be
+                // scrolled, so the count always has something to
+                // divide and the control is always worth offering.
+                showColumns = true,
+                onLineHeightChanged = onLineHeight,
+                onPageMarginsChanged = onPageMargins,
+                onColumnModeChanged = onColumnMode,
+            )
+            ReadingFooterModeDropdown(
+                selected = prefs.footerMode,
+                onSelected = onFooterMode,
+            )
+            ReadingPageTurnAnimationToggle(
+                enabled = prefs.pageTurnAnimation,
+                onChanged = onPageTurnAnimation,
+            )
         }
     }
 }
